==> 09_Figure_S2.Rmd <==
---
title: "A novel role for alkyladenine DNA glycosylase in regulating alkylation-induced ER stress"
subtitle: "Figure S2"
author: "L Milano, CF Charlier, R Andreguetti, T Cox, E Healing, MP Thomé, RM Elliott, JY Masson, LD Samson, JY Masson, G Lenz, JAP Henriques, A Nohturfft and LB Meira"
output:
  html_notebook:
    toc: TRUE
    toc_float: TRUE
    toc_depth: 3
    code_folding: "hide"
    number_sections: FALSE
    theme: "readable"
    highlight: "tango"
    fig_caption: TRUE
    css: "./source/styles.css"
---

```{js}
function myFunction(id) {
  var x = document.getElementById(id);
  if (x.style.display === 'none') {
    x.style.display = 'block';
  } else {
    x.style.display = 'none';
  }
}
```


<div class="comments">
***Introduction:***  
Gene-set enrichment analysis showed that genes induced by MMS in wild-type liver are enriched for many genesets induced by drug treatments. Many of these drugs are known to induce ER stress!  

The finding supports the conclusion that MMS induces ER stress in mouse liver.
  
***Objective:***  
Plot a clustered binary heatmap of genes induced by MMS (this study) and by other drugs known to induce ER stress.  
</div>

<div class='comments'>
***Note***: This analysis requires input data generated by an earlier script.
</div>

## Setup  

* Clear memory and console  
* Load packages  
* Get current script name  

```{r CLEAR MEMORY AND PACKAGES AND CONSOLE, results="hide"}
# Clear memory
rm(list=setdiff(ls(all=TRUE), c(".Random.seed")))
# Clear console
cat("\014")
```


```{r include=FALSE}
library(magrittr)
library(dplyr)
library(openxlsx)
library(Hmisc)
```

 
```{r class.output="txt_output"}
this.script <- rstudioapi::getActiveDocumentContext()$path %>% basename
cat(cat("Script:", this.script))
```


## Analyse enrichment results for genes induced by MMS in wild-type liver    
### Load enrichment data  
```{r class.output="txt_output"}
# list.files()
enriched.xlsx.file <- "Supplemental_Data_2.xlsx"
the.sheet <- "wt.mms.vs.wt.ctrl_up"
stopifnot(file.exists(enriched.xlsx.file))
enriched.wt.up <- openxlsx::read.xlsx(xlsxFile = enriched.xlsx.file,
                                      sheet=the.sheet, colNames = TRUE, startRow = 5)
cat("File read: ", enriched.xlsx.file, " (Worksheet '", "wt.mms.vs.wt.ctrl_up", "')", sep="")
rm(enriched.xlsx.file)
```


### Select enriched library terms related to drug regulation  
```{r results="asis"}
drug.libs <- grep("drug", enriched.wt.up$Library, ignore.case=TRUE, value=TRUE) %>% unique %>% sort
cat("**Matching enrichr libraries:**  \n\n", paste0("* ", drug.libs, "  \n"), sep="")
```
### Filter on terms related to up regulation by drugs  
```{r class.output="txt_output"}
drug.libs.up.df <- enriched.wt.up %>% 
  filter(Library %in% drug.libs) %>% 
  filter(grepl("up$", Library, ignore.case=T) | grepl("up$", Term, ignore.case=T))
cat("Number of matching library terms:", nrow(drug.libs.up.df))
```

```{r}
DT::datatable(drug.libs.up.df[,1:2], caption="Terms related to upregulation by drugs", options=list(pageLength=5))
```

## Drugs known to induce ER stress  
### Select ER-stress-inducing drugs  
```{r}
er.stress.inducers <- c("Acetaminophen", "Bisphenol A", "Cadmium", "Carboplatin",
                        "PBDE 47", "PBDE 85", "Palmitic acid", "Progesterone",
                        "Tocotrienol", "Ursodeoxycholic acid"
)
```

```{r}
# er.stress.inducers <- c("Acetaminophen", "Arsenite", "Azacitidine", "Bisphenol A",
#                         "Cadmium", "Carbon Tetrachloride", "Carboplatin", "Chlorhexidine",
#                         "Colchicine", "Dexamethasone", "Geldanamycin", "Ionomycin",
#                         "MLN4924", "Niclosamide", "PBDE 47", "PBDE 85", "Palmitic acid",
#                         "Progesterone", "Spiperone", "Sulindac", "Tanespimycin",
#                         "Thioridazine", "Tocotrienol", "Ursodeoxycholic acid")
```


<div class="comments">
A literature study was performed on the drugs listed in the `r nrow(drug.libs.up.df)` enrichr library terms above. This search identified `r length(er.stress.inducers)` drugs from the list that are known to induce ER stress:
</div>

```{r results="asis", class.output="txt_output"}
cat("**ER stress inducers:**  \n\n",
    paste0("* ", er.stress.inducers, "  \n"), sep="")
```


### Filter enrichment data on library terms related to ER stress-inducing drugs  
```{r}
# Check that all drugs are covered in the 'Term' column:
x <- sapply(er.stress.inducers, function(x) {
  any(grepl(x, drug.libs.up.df$Term, ignore.case=T))
})
# sum(x)
# dput(er.stress.inducers[x])
stopifnot(all(x))
rm(x)
```


```{r class.output="txt_output"}
# Generate a search string to be used with grep below:  
search.str <- paste(er.stress.inducers, collapse="|")
# Subsetting:
drug.libs.up.stress.df <- drug.libs.up.df %>% 
  dplyr::filter(grepl(search.str, Term, ignore.case=T))
rm(search.str)
cat("Number of library terms selected:", nrow(drug.libs.up.stress.df))
```

```{r}
drug.libs.up.stress.df %>% 
  dplyr::select(Library, Term, Adj.P.val) %>% 
  dplyr::arrange(Adj.P.val) %>% 
  dplyr::mutate(Adj.P.val = signif(Adj.P.val, digits=3)) %>% 
  DT::datatable(., options = list(pageLength=5))
```


Generate a list object that stores overexpressed genes for each of the ER stress inducing drugs:    
```{r}
er.stress.inducers
get.short.title <- function(x) {
  y <- gsub("Bisphenol A", "Bisphenol_A", x, ignore.case = T)
  y <- gsub("gamma-Tocotrienol", "Tocotrienol", y, ignore.case = T)
  y <- gsub("PBDE 47", "PBDE_47", y, ignore.case = T)
  y <- gsub("PBDE 85", "PBDE_85", y, ignore.case = T)
  y <- gsub("ursodeoxycholic acid", "Ursodeoxycholate", y, ignore.case = T)
  y <- gsub("palmitic acid", "Palmitate", y, ignore.case = T)
  y <- gsub("CADMIUM", "Cadmium", y, ignore.case = T)
  z <- strsplit(y, split="[ -]")[[1]][1]
  Hmisc::capitalize(z)
}
```

```{r}
# names(drug.libs.up.stress.df)
drugs.gene.list <- lapply(seq_len(nrow(drug.libs.up.stress.df)), function(i, er.stress.inducers) {
  # i <- 1
  lib.term <- unlist(drug.libs.up.stress.df$Term[i])
  # gene.set
  lib.term.short <- get.short.title(lib.term)
  # gene.set.short
  list(
    Term = lib.term,
    Term.short = lib.term.short,
    Genes = strsplit(drug.libs.up.stress.df$Genes[i], split=",")[[1]] 
  )
}, er.stress.inducers=er.stress.inducers)
```

```{r results="asis"}
x <- paste(utils::capture.output(str(drugs.gene.list)), collapse="<br>\n", sep="")
details::details(x, summary="Show summary of gene list object", lang=NULL)
rm(x)
```


Aggregate gene list by drug  
```{r}
combine.genes <- function(vec) {
  a <- strsplit(vec, split=";")[[1]]
  b <- sort(unique(a))
  c <- paste(b, collapse=";")
  c
}
```

```{r}
# unique(sapply(drugs.gene.list, class)) # list
gene.list.aggr <- lapply(drugs.gene.list, function(lst.i) {
  a <- data.frame(Drug = lst.i$Term.short,
                  Genes = lst.i$Genes,
                  stringsAsFactors = FALSE)
  a
}) %>% 
  data.table::rbindlist(.) %>% 
  dplyr::group_by(Drug) %>%
  dplyr::summarise(Genes = paste(Genes, collapse=";")) %>%
  dplyr::ungroup()

gene.list.aggr$Genes <- sapply(gene.list.aggr$Genes, combine.genes)
# DT::datatable(gene.list.aggr)
```


Convert back to list object  
```{r}
aggr.gene.list <- lapply(seq_len(nrow(gene.list.aggr)), function(i) {
    strsplit(gene.list.aggr$Genes[i], split=";")[[1]]
}) %>%
  set_names(gene.list.aggr$Drug)
```

```{r results="asis"}
x <- paste(utils::capture.output(str(aggr.gene.list)), collapse="<br>\n", sep="")
details::details(x, summary="Show summary of aggregated gene list object", lang=NULL)
rm(x)
```


```{r class.output="txt_output"}
# How many different genes in total?  
uniq.genes <- unlist(aggr.gene.list) %>% unique %>% sort
cat("Total number of genes covered:", length(uniq.genes))
```

### Make a matrix:  

* rows = genes
* columns Drugs  
* values: true/false whether gene is present in gene-set  

```{r}
mx <- lapply(uniq.genes, function(g) {
  sapply(aggr.gene.list, function(genes.v, g) {
    as.numeric(g %in% genes.v)
  }, g=g)
}) %>% do.call(rbind, .) %>% 
  magrittr::set_rownames(uniq.genes) %>% 
  magrittr::set_colnames(names(aggr.gene.list))
```

Print excerpt of matrix (6 rows, 6 columns):
```{r}
mx[1:4, 1:6]
```


## Heatmap: all genes
```{r}
gplots::heatmap.2(x=t(mx), distfun= function(c) dist(c, method="manhattan"), trace="none", key=F,
                  col=c("grey90", "black"))
```


## Reduce number of genes covered  

<div class="comments">
To generate a more readable graphic select only those genes that appear at least three times.
</div>

Get counts for how often each gene was covered:  
```{r}
gene.counts.df <- table(unlist(aggr.gene.list)) %>% 
  as.data.frame(., stringsAsFactors=FALSE) %>% 
  set_colnames(c("Gene", "Count")) %>% 
  dplyr::arrange(desc(Count))

table(gene.counts.df$Count)
```

```{r class.output="txt_output"}
genes.select <- subset(gene.counts.df , Count >= 3) %>% .[,"Gene"]
cat("Number of genes selected:", length(genes.select))
```

Generate new matrix:  
```{r}
mx.select <- mx[genes.select, ]
```


```{r}
plot.it <- function(mtrx) {
  gplots::heatmap.2(x=mtrx, distfun= function(c) dist(c, method="binary"), trace="none", key=F,
                  col=c("grey90", "black"))
}
plot.it(mx.select)
```

```{r}
pdf.file <- "Figure_S2.pdf"
pdf(file=pdf.file, paper="a4", width=7, height=10, pointsize = 6)
plot.it(mx.select)
dev.off()
```

```{r class.output="txt_output"}
cat("File saved:", pdf.file)
rm(pdf.file)
```


<hr>

## Session info  

<button class="button" onclick="myFunction('DIV_1')">Show/hide session info</button>
<div id="DIV_1" class="div_default_hide">

```{r SESSION INFO DATE, results="asis"}
cat("Date:", format(Sys.time(), "%a %d-%b-%Y %H:%M:%S"), "<br>\n")
```

```{r print_session_info, R.options=list(width=70)}
devtools::session_info()
```
</div>
  
```{js}
var divsToHide = document.getElementsByClassName("div_default_hide");
for(var i = 0; i < divsToHide.length; i++)
{
  divsToHide[i].style.display = 'none';
}
```
